Update Temperature-Display-LED-Strip.ino
I have added the implementation for an analog temperature sensor. The digital sensor is still present in the code because it is used for calibrating the analog sensor with the temperature from the digital sensor, and it is also used for obtaining humidity data.

Important: This implementation has not been tested yet.